Move some slow tests from `test_lstar.py` into `nocover/`
