+++ hypothesis-python/tests/conjecture/test_lstar.py
@@ -204,25 +204,6 @@ def test_learning_large_dfa():
         assert i == int.from_bytes(s, "big")
 
 
-@st.composite
-def byte_order(draw):
-    ls = draw(st.permutations(range(256)))
-    n = draw(st.integers(0, len(ls)))
-    return ls[:n]
-
-
-@example({0}, [1])
-@given(st.sets(st.integers(0, 255)), byte_order())
-def test_learning_always_changes_generation(chars, order):
-    learner = LStar(lambda s: len(s) == 1 and s[0] in chars)
-    for c in order:
-        prev = learner.generation
-        s = bytes([c])
-        if learner.dfa.matches(s) != learner.member(s):
-            learner.learn(s)
-            assert learner.generation > prev
-
-
 def varint_predicate(b):
     if not b:
         return False
